# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 86.88 B&B — REST API
  version: 1.0.0
  description: |
    Backend API for the 86.88 民宿 (B&B) platform.

    Serves as the bridge between the PostgreSQL database and the frontend apps:
    - `apps/website` — public-facing Next.js site (8688bnb.com)
    - `apps/admin` — admin dashboard (admin.8688bnb.com)

    ## Authentication
    Admin authentication uses JWT tokens stored in `__Host-8688_session` HttpOnly cookies.
    All state-changing requests require a CSRF token via the `X-CSRF-Token` header.

    ## Rate Limiting
    - Auth endpoints: 20 requests / 15 minutes
    - General API: 200 requests / 15 minutes
    - Public reads: 500 requests / 15 minutes
    - Webhook ingress: 100 requests / 15 minutes

    Rate limit info is returned in response headers:
    - `X-RateLimit-Limit` — max requests in window
    - `X-RateLimit-Remaining` — requests left
    - `X-RateLimit-Reset` — window reset time (Unix epoch)

  contact:
    name: 8688 BnB Dev
    email: dev@8688bnb.com
  license:
    name: Private
    identifier: UNLICENSED
servers:
  - url: http://localhost:3333/api/v1
    description: Local development
  - url: https://api.8688bnb.com/api/v1
    description: Production (via Cloudflare Tunnel)
...
security: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-8688_session
      description: |
        JWT stored in a hardened HttpOnly cookie.
        Set automatically on successful login.
        Cookie attributes: HttpOnly, Secure, SameSite=Strict, Path=/
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token required for all state-changing requests (POST, PUT, DELETE).
        Obtained from the login response or `GET /auth/csrf-token`.
    webhookApiKey:
      type: apiKey
      in: header
      name: X-Webhook-Key
      description: |
        API key for authenticating incoming OTA channel manager webhooks.
        Configured in the WEBHOOK_API_KEY environment variable.

  # ─── Reusable Schemas ────────────────────────────────────────────────────────
  schemas:
...
    MediaItem:
      type: object
      description: |
        A single image/media file.
        Media is organized by "target" — which page/section it belongs to.
        Target examples: homepage_hero, homepage_cats, homepage_bnb,
        about, rooms_overview, room_r1, room_r2, ..., location, gallery, etc.
      properties:
        id:
          type: integer
          example: 1
        target:
          type: string
          example: homepage_hero
          description: |
            Which page/section this media belongs to.
            Convention follows the site structure:
            - homepage_hero — Hero carousel images
            - homepage_8688 — 8688 section images
            - homepage_cats — Cats section images
            - homepage_bnb — BnB section images
            - about — About Us page
            - rooms_overview — Rooms overview (stay) page thumbnails
            - room_{slug} — Per-room gallery (e.g. room_r1, room_r2)
            - booking_info — Booking info / regulation page
            - location — Location page
            - gallery — General photo gallery
            - brand — Logo and brand assets
        filename_original:
          type: string
          example: room1-view.jpg
          description: Original uploaded filename (stored in DB only, not used for storage)
        filename_stored:
          type: string
          example: a3f8c2e1-4d5b-6789-0abc-def012345678.webp
          description: UUID-based filename on disk
        url:
          type: string
          format: uri
          example: /uploads/images/a3f8c2e1-4d5b-6789-0abc-def012345678.webp
          description: Public URL path to serve the image
        mime_type:
          type: string
          enum: [image/jpeg, image/png, image/webp]
          example: image/webp
        size_bytes:
          type: integer
          example: 245760
        width:
          type: integer
          nullable: true
          example: 1920
        height:
          type: integer
          nullable: true
          example: 1080
        alt_text:
          type: string
          nullable: true
          example: 浪漫情懷房間景觀
          description: Alt text for accessibility and SEO
        sort_order:
          type: integer
          example: 1
          description: Display order within the target group
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - target
        - filename_original
        - filename_stored
        - url
        - mime_type
        - size_bytes
        - sort_order
        - created_at
        - updated_at
    MediaUploadResponse:
      type: object
      properties:
        media:
          $ref: '#/components/schemas/MediaItem'
        message:
          type: string
          example: 圖片上傳成功
      required: [media, message]
    MediaReorderRequest:
      type: object
      properties:
        image_ids:
          type: array
          items:
            type: integer
          description: Ordered list of media item IDs (new display order)
          minItems: 1
          maxItems: 50
      required: [image_ids]

    # ── Pages CMS ──────────────────────────────────────────────────────────────
    Page:
      type: object
      properties:
        id:
          type: integer
          example: 1
        slug:
          type: string
          example: about
          description: URL path segment
        title_zh:
          type: string
          example: 關於我們
        title_en:
          type: string
          nullable: true
          example: About Us
        content_html:
          type: string
          description: Sanitized HTML content
          example: <h2>86.88 民宿</h2><p>位於宜蘭三星...</p>
        meta:
          type: object
          nullable: true
          additionalProperties: true
          description: SEO metadata, Open Graph tags, etc.
          example:
            seo_title: 關於我們 — 86.88 民宿
            seo_description: 宜蘭三星歐風民宿，貓咪陪伴療癒人心
        published:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, slug, title_zh, published, created_at, updated_at]
    PageUpdateRequest:
      type: object
      properties:
        title_zh:
          type: string
          minLength: 1
          maxLength: 200
        title_en:
          type: string
          maxLength: 200
        content_html:
          type: string
          maxLength: 50000
          description: HTML content — will be sanitized server-side (DOMPurify)
        meta:
          type: object
          additionalProperties: true
        published:
          type: boolean
      minProperties: 1

    # ── News / Announcements (最新消息) ────────────────────────────────────────
    NewsItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 中秋連假加開房間
        content:
          type: string
          example: 中秋連假期間(9/15-9/17)加開一間房...
        published_at:
          type: string
          format: date
          nullable: true
          example: '2026-09-01'
        visible:
          type: boolean
          example: true
        pinned:
          type: boolean
          example: false
          description: Pinned news items always show at the top
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, title, content, visible, pinned, created_at, updated_at]
    NewsCreateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: 中秋連假加開房間
        content:
          type: string
          minLength: 1
          maxLength: 5000
        published_at:
          type: string
          format: date
          description: Schedule future publish date, or omit for immediate
        visible:
          type: boolean
          default: true
        pinned:
          type: boolean
          default: false
      required: [title, content]
    NewsUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
          maxLength: 5000
        published_at:
          type: string
          format: date
          nullable: true
        visible:
          type: boolean
        pinned:
          type: boolean
      minProperties: 1

    # ── OTA Webhooks (Unified Channel Manager) ─────────────────────────────────
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          example: 47
        total_pages:
          type: integer
          example: 3
      required: [page, per_page, total, total_pages]
...
    ApiError:
      type: object
      description: Standard error response
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
              description: |
                Machine-readable error code. Examples:
                - AUTH_INVALID_CREDENTIALS
                - AUTH_SESSION_EXPIRED
                - AUTH_CSRF_INVALID
                - VALIDATION_ERROR
                - NOT_FOUND
                - RATE_LIMITED
                - CONFLICT (e.g. double-booking)
                - INTERNAL_ERROR
                - UPLOAD_FILE_TOO_LARGE
                - UPLOAD_INVALID_TYPE
            message:
              type: string
              example: 驗證失敗
              description: Human-readable error message (Chinese)
            message_en:
              type: string
              example: Validation failed
              description: English fallback message
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: check_in
                  message:
                    type: string
                    example: 入住日期不得早於今天
              description: Per-field validation errors (if applicable)
          required: [code, message]
      required: [success, error]
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
...
    NewsIdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: News item ID
    MediaIdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Media item ID
...
  responses:
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: AUTH_SESSION_EXPIRED
              message: 登入已過期，請重新登入
              message_en: Session expired, please log in again
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: 找不到資源
              message_en: Resource not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: 資料驗證失敗
              message_en: Validation failed
              details:
                - field: check_in
                  message: 入住日期不得早於今天
...
paths:
...
  /media:
    get:
      tags: [Media]
      operationId: listMedia
      summary: List media items
      description: |
        List media items, optionally filtered by target (page/section).
        Public for serving images, admin for management.

        Target values map to website sections:
        - `homepage_hero` — Hero carousel
        - `homepage_8688`, `homepage_cats`, `homepage_bnb` — Homepage sections
        - `about` — About page
        - `rooms_overview` — Rooms overview thumbnails
        - `room_r1`, `room_r2`, ... — Per-room galleries
        - `booking_info` — Booking info page
        - `location` — Location page
        - `gallery` — General photo gallery
        - `brand` — Logo and brand assets
      parameters:
        - name: target
          in: query
          schema:
            type: string
          description: Filter by target (page/section slug)
          example: homepage_hero
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Media list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaItem'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                required: [success, data]
  /media/targets:
    get:
      tags: [Media]
      operationId: listMediaTargets
      summary: List available media targets
      description: |
        Returns all defined media targets (page/section identifiers) with
        their display names and current image count. Useful for the admin
        media manager UI to show which sections have images.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Media target list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        target:
                          type: string
                          example: homepage_hero
                        label_zh:
                          type: string
                          example: 首頁輪播
                        label_en:
                          type: string
                          example: Homepage Hero Carousel
                        image_count:
                          type: integer
                          example: 4
                      required: [target, label_zh, image_count]
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
  /media/upload:
    post:
      tags: [Media]
      operationId: uploadMedia
      summary: Upload image to a page/section
      description: |
        Upload an image file for a specific page/section target.

        Security:
        - Allowed types: JPEG, PNG, WebP (validated by magic bytes, not just extension)
        - Max file size: 10MB
        - Filename is replaced with a UUID
        - Original filename stored in DB only

        Target must be a valid section identifier (see GET /media/targets).
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, or WebP, max 10MB)
                target:
                  type: string
                  example: homepage_hero
                  description: Which page/section this image belongs to
                alt_text:
                  type: string
                  maxLength: 200
                  example: 民宿外觀夜景
                  description: Alt text for accessibility/SEO
                sort_order:
                  type: integer
                  minimum: 0
                  description: Display order within the target group
              required: [file, target]
      responses:
        '201':
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadResponse'
        '400':
          description: Invalid file type or missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: UPLOAD_INVALID_TYPE
                  message: 只允許 JPEG、PNG、WebP 格式的圖片
                  message_en: Only JPEG, PNG, and WebP images are allowed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: UPLOAD_FILE_TOO_LARGE
                  message: 檔案大小超過 10MB 上限
                  message_en: File size exceeds 10MB limit
  /media/{id}:
    parameters:
      - $ref: '#/components/parameters/MediaIdParam'

    put:
      tags: [Media]
      operationId: updateMedia
      summary: Update media metadata
      description: Update alt text, sort order, or target of a media item.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alt_text:
                  type: string
                  maxLength: 200
                target:
                  type: string
                  description: Move image to a different page/section
                sort_order:
                  type: integer
                  minimum: 0
              minProperties: 1
      responses:
        '200':
          description: Media updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/MediaItem'
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Media]
      operationId: deleteMedia
      summary: Delete media item
      description: |
        Delete a media item. Removes both the DB record and the file from disk.
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        '200':
          description: Media deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 圖片已刪除
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /media/reorder:
    put:
      tags: [Media]
      operationId: reorderMedia
      summary: Reorder media items within a target
      description: |
        Set the display order of media items within a target group.
        Provide an ordered array of media IDs — they will be assigned
        sort_order values 0, 1, 2, ... in that order.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaReorderRequest'
      responses:
        '200':
          description: Media reordered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 圖片排序已更新
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ═══════════════════════════════════════════════════════════════════════════
  # PAGES (CMS)
  # ═══════════════════════════════════════════════════════════════════════════
  /pages:
    get:
      tags: [Pages]
      operationId: listPages
      summary: List all pages
      description: |
        Public endpoint — returns published pages.
        Admin users can see drafts with `include_drafts=true`.
      parameters:
        - name: include_drafts
          in: query
          schema:
            type: boolean
            default: false
          description: Include unpublished draft pages (admin only)
      responses:
        '200':
          description: Page list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Page'
                required: [success, data]
  /pages/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        description: Page slug (e.g. about, booking-info, location)

    get:
      tags: [Pages]
      operationId: getPage
      summary: Get page by slug
      description: Public endpoint — returns page content.
      responses:
        '200':
          description: Page content
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Page'
                required: [success, data]
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Pages]
      operationId: updatePage
      summary: Update page content (CMS)
      description: |
        Update page title, content, meta, or publish status.
        HTML content is sanitized server-side using DOMPurify.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageUpdateRequest'
      responses:
        '200':
          description: Page updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Page'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # NEWS (最新消息)
  # ═══════════════════════════════════════════════════════════════════════════
  /news:
    get:
      tags: [News]
      operationId: listNews
      summary: List news items
      description: |
        Public endpoint — returns visible, published news items.
        Sorted by pinned first, then by published_at descending.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: include_hidden
          in: query
          schema:
            type: boolean
            default: false
          description: Include hidden news items (admin only)
      responses:
        '200':
          description: News list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NewsItem'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                required: [success, data]

    post:
      tags: [News]
      operationId: createNews
      summary: Create news item
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsCreateRequest'
      responses:
        '201':
          description: News created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/NewsItem'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /news/{id}:
    parameters:
      - $ref: '#/components/parameters/NewsIdParam'

    get:
      tags: [News]
      operationId: getNewsItem
      summary: Get news item
      responses:
        '200':
          description: News detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/NewsItem'
                required: [success, data]
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [News]
      operationId: updateNews
      summary: Update news item
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsUpdateRequest'
      responses:
        '200':
          description: News updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/NewsItem'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [News]
      operationId: deleteNews
      summary: Delete news item
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        '200':
          description: News deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 消息已刪除
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # DASHBOARD
  # ═══════════════════════════════════════════════════════════════════════════
...